pyproject.toml
[project]
name = "book-store-psql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "asyncpg==0.30.0",
    "bcrypt==4.3.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "confluent-kafka>=2.11.0",
    "cryptography==44.0.2",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "greenlet==3.2.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jose==1.0.0",
    "jwt==1.3.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "motor==3.7.0",
    "passlib==1.7.4",
    "pip==25.0.1",
    "pyasn1==0.4.8",
    "pycparser==2.22",
    "pydantic==2.11.2",
    "pydantic-core==2.33.1",
    "pydantic-settings>=2.10.1",
    "pygments==2.19.1",
    "pymongo==4.11.3",
    "python-dotenv==1.1.0",
    "python-jose==3.4.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests>=2.32.4",
    "rich==14.0.0",
    "rich-toolkit==0.14.1",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "typer==0.15.2",
    "typing-extensions==4.13.1",
    "typing-inspection==0.4.0",
    "uvicorn==0.34.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
]
